fru: added command line parameters for offset
Instead order to use fru-tools to interact with multiple
types of EEPROM, command line parameters (-I & -O)
were added for the input file offset and output file
offset. This way, fru-dump can operate in the desired
area and can be used with a greater variety of
EEPROM files (such as RPI EEPROM).

Signed-off-by: Ramona Alexandra Nechita <ramona.nechita@analog.com>